+++ metastor/managedObjectLoader/src/main/java/org/finra/herd/metastore/managed/format/ColumnDef.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
 **/
-package org.finra.herd.metastore.managed.hive;
+package org.finra.herd.metastore.managed.format;
 
 import lombok.*;
