+++ metastor/managedObjectLoader/src/main/java/org/finra/herd/metastore/managed/format/FormatChange.java
@@ -18,8 +18,6 @@ package org.finra.herd.metastore.managed.format;
 import com.google.common.collect.Lists;
 import lombok.*;
 import org.apache.commons.math3.util.Pair;
-import org.finra.herd.metastore.managed.hive.ClusteredDef;
-import org.finra.herd.metastore.managed.hive.ColumnDef;
 
 import java.util.List;
