+++ metastor/managedObjectLoader/src/main/java/org/finra/herd/metastore/managed/hive/HiveClientImpl.java
@@ -19,6 +19,8 @@ import com.google.common.collect.Lists;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.finra.herd.metastore.managed.format.ClusteredDef;
+import org.finra.herd.metastore.managed.format.ColumnDef;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
