+++ metastor/managedObjectLoader/src/test/java/org/finra/herd/metastore/managed/NotificationSenderTest.java
@@ -18,7 +18,7 @@ package org.finra.herd.metastore.managed;
 import com.google.common.collect.Lists;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.math3.util.Pair;
-import org.finra.herd.metastore.managed.hive.ColumnDef;
+import org.finra.herd.metastore.managed.format.ColumnDef;
 import org.finra.herd.metastore.managed.format.FormatChange;
 import org.finra.herd.metastore.managed.hive.HiveTableSchema;
 import org.junit.Test;
